rename some things and TODO comments
