+++ src/main/java/org/cancogenvirusseq/muse/components/SongScoreClient.java
@@ -8,9 +8,9 @@ import lombok.Builder;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-import org.cancogenvirusseq.muse.model.song_score_client.AnalysisFileResponse;
-import org.cancogenvirusseq.muse.model.song_score_client.ScoreFileSpec;
-import org.cancogenvirusseq.muse.model.song_score_client.SubmitResponse;
+import org.cancogenvirusseq.muse.model.song_score.AnalysisFileResponse;
+import org.cancogenvirusseq.muse.model.song_score.ScoreFileSpec;
+import org.cancogenvirusseq.muse.model.song_score.SubmitResponse;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
